backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
httpx
python-dotenv
python-multipart
